refactor: move types used in booking to BookingTypes
